Move stray files to utils directory
